common/bitutil: optimize TestBytes
